refactor(mark, monument): unify validation state handling by introducing shared ValidationState enum and converter
