+++ mark/src/main/java/pt/estga/mark/entities/Mark.java
@@ -2,13 +2,13 @@ package pt.estga.mark.entities;
 
 import jakarta.persistence.*;
 import lombok.*;
-import pt.estga.mark.converters.ValidationStateConverter;
+import pt.estga.shared.converters.ValidationStateConverter;
 import pt.estga.shared.entities.BaseEntity;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import pt.estga.mark.enums.MarkValidationState;
+import pt.estga.shared.enums.ValidationState;
 
 @Entity
 @NoArgsConstructor
@@ -40,6 +40,6 @@ public class Mark extends BaseEntity {
     @Builder.Default
     @Convert(converter = ValidationStateConverter.class)
     @Column(nullable = false)
-    private MarkValidationState validationState = MarkValidationState.PROVISIONAL;
+    private ValidationState validationState = ValidationState.PROVISIONAL;
 
 }
